Refactor & consolidate inputs and fields
Solves #85
