+++ src/Ui/Forms/Fields/CheckboxField/CheckboxField.hbs
@@ -1,1 +1,1 @@
-<kirby-checkbox-field text=""></kirby-checkbox-field>
+<kirby-checkbox-field text="This is a checkbox" />
