+++ src/Ui/Forms/Fields/CheckboxField/CheckboxField.vue
@@ -1,36 +1,32 @@
 <template>
   <kirby-field class="kirby-checkbox-field" v-bind="$props">
-    <kirby-checkbox-input v-model="data" :label="text" />
+    <kirby-checkbox-input :text="text" v-model="data" />
   </kirby-field>
 </template>
 
 <script>
 
-import Props from '../Field.props.js';
+import Field from '../Field.mixin.js';
 
 export default {
-  mixins: [Props],
+  mixins: [Field],
   props: {
-    label: {
-      type: String
-    },
-    text: {
-      type: String
-    },
-    name: {
-      type: String,
-      default: 'check'
-    },
     value: {
       type: Boolean,
       default: false
+    },
+    name: {
+      default: 'check'
+    },
+    text: {
+      type: String
     }
   }
 }
 
 </script>
 
-<style>
+<style lang="scss">
 
 .kirby-checkbox-field .kirby-input-content {
   padding: .65rem;
