+++ src/Ui/Forms/Fields/CheckboxesField/CheckboxesField.hbs
@@ -3,4 +3,4 @@
     { value: 'b', text: 'Option B' },
     { value: 'c', text: 'Option C' },
     { value: 'd', text: 'Option D' }
-]" :value="['b', 'd']"><kirby-checkboxes-field>
+]" :value="['b', 'd']" />
