+++ src/Ui/Forms/Fields/CheckboxesField/CheckboxesField.vue
@@ -1,26 +1,42 @@
 <template>
   <kirby-field class="kirby-checkboxes-field" v-bind="$props">
-    <kirby-checkboxes-input v-model="data" v-bind="$props" />
+    <kirby-checkbox-input
+    v-for="(option, index) in options"
+    :key="index"
+    :value="data.indexOf(option.value) !== -1"
+    :text="option.text"
+    @input="input($event, index)"
+    />
   </kirby-field>
 </template>
 
 <script>
 
-import Props from '../Field.props.js';
+import Field from '../Field.mixin.js';
 
 export default {
-   mixins: [Props],
+  mixins: [Field],
   props: {
-    options: {
-      type: Array
-    },
-    name: {
-      type: String,
-      default: 'checks'
-    },
     value: {
       type: Array,
       default: []
+    },
+    name: {
+      default: 'checks'
+    },
+    options: {
+      type: Array,
+      default: []
+    }
+  },
+  methods: {
+    input (value, index) {
+      let option = this.options[index].value;
+      if (value === true) {
+        this.data.push(option)
+      } else {
+        this.data = this.data.filter(item => item !== option)
+      }
     }
   }
 }
@@ -31,6 +47,10 @@ export default {
 
 .kirby-checkboxes-field .kirby-input-content {
   padding: .65rem;
+}
+
+.kirby-checkboxes-field > * + * {
+  margin-top: 4px;
 }
 
 </style>
