+++ src/Ui/Forms/Fields/DateField/DateField--with-Value.hbs
@@ -1,0 +1,1 @@
+<kirby-date-field value="1989-01-29" />
