+++ src/Ui/Forms/Fields/DateField/DateField.hbs
@@ -1,1 +1,1 @@
-<kirby-date-field><kirby-date-field>
+<kirby-date-field />
