+++ src/Ui/Forms/Fields/EmailField/EmailField.hbs
@@ -1,1 +1,1 @@
-<kirby-email-field></kirby-email-field>
+<kirby-email-field />
