+++ src/Ui/Forms/Fields/HeadlineField/HeadlineField.hbs
@@ -1,1 +1,1 @@
-<kirby-headline-field label="This is a headline"></kirby-headline-field>
+<kirby-headline-field label="This is a headline" />
