+++ src/Ui/Forms/Fields/InfoField/InfoField.hbs
@@ -1,3 +1,1 @@
-<kirby-info-field>
-    This is an informational text.
-</kirby-info-field>
+<kirby-info-field text="This is an informational text." />
