+++ src/Ui/Forms/Fields/NumberField/NumberField.hbs
@@ -1,1 +1,1 @@
-<kirby-number-field><kirby-number-field>
+<kirby-number-field />
