+++ src/Ui/Forms/Fields/PasswordField/PasswordField.hbs
@@ -1,1 +1,1 @@
-<kirby-password-field></kirby-password-field>
+<kirby-password-field />
