+++ src/Ui/Forms/Fields/TagsField/TagsField--with-autocomplete.hbs
@@ -1,4 +1,4 @@
-<kirby-tags-input :autocomplete="{
+<kirby-tags-field :autocomplete="{
   url: '/mock/data/countries.json',
   map: {value: 'name', text: 'name'}
 }" :sortable="false" />
