+++ src/Ui/Forms/Fields/TagsField/TagsField--with-custom-separator.hbs
@@ -1,0 +1,1 @@
+<kirby-tags-field :value="'design:photography:architecture'" :separator="':'" />
