+++ src/Ui/Forms/Fields/TagsField/TagsField.hbs
@@ -1,1 +1,1 @@
-<kirby-tags-field><kirby-tags-field>
+<kirby-tags-field />
