+++ src/Ui/Forms/Fields/TextField/TextField.hbs
@@ -4,5 +4,4 @@
   icon="edit"
   placeholder="Write something there…"
   help="This is some nice help"
-  :required="true">
-</kirby-text-field>
+  :required="true" />
