+++ src/Ui/Forms/Fields/TextField/TextField.props.js
@@ -1,9 +1,9 @@
 
-import Input from '../../Inputs/TextInput/TextInput.props.js';
+// import Input from '../../Inputs/TextInput/TextInput.props.js';
 import Field from '../Field.props.js';
 
 export default {
-  mixins: [Input, Field],
+  mixins: [Field],
   props: {
   }
 }
