+++ src/Ui/Forms/Fields/TextareaField/TextareaField.hbs
@@ -1,1 +1,1 @@
-<kirby-textarea-field><kirby-textarea-field>
+<kirby-textarea-field />
