+++ src/Ui/Forms/Fields/ToggleField/ToggleField--static.hbs
@@ -1,1 +1,1 @@
-<kirby-toggle-field text="I am resistant to change."><kirby-toggle-field>
+<kirby-toggle-field text="I am resistant to change." />
