+++ src/Ui/Forms/Fields/ToggleField/ToggleField.hbs
@@ -1,1 +1,1 @@
-<kirby-toggle-field on="It's on" off="So off!"><kirby-toggle-field>
+<kirby-toggle-field on="It's on" off="So off!" />
