+++ src/Ui/Forms/Fields/UrlField/UrlField.hbs
@@ -1,1 +1,1 @@
-<kirby-url-field></kirby-url-field>
+<kirby-url-field />
