+++ src/Ui/Forms/Inputs/CheckboxInput/CheckboxInput.hbs
@@ -1,1 +1,1 @@
-<kirby-checkbox-input label="This is a checkbox" />
+<kirby-checkbox-input />
