--- src/Ui/Forms/Inputs/CheckboxesInput/CheckboxesInput.hbs
@@ -1,6 +1,0 @@
-<kirby-checkboxes-input :options="[
-    { value: 'a', text: 'Option A' },
-    { value: 'b', text: 'Option B' },
-    { value: 'c', text: 'Option C' },
-    { value: 'd', text: 'Option D' }
-]" :value="['b', 'd']" />
