--- src/Ui/Forms/Inputs/DateInput/DateInput.hbs
@@ -1,1 +1,0 @@
-<kirby-date-input></kirby-date-input>
