+++ src/Ui/Forms/Inputs/Input.mixin.js
@@ -1,0 +1,14 @@
+
+export default {
+  props: ['value'],
+  data () {
+    return {
+      data: this.value
+    }
+  },
+  watch: {
+    data () {
+      this.$emit('input', this.data)
+    }
+  }
+}
