--- src/Ui/Forms/Inputs/TagsInput/TagsInput--with-custom-separator.hbs
@@ -1,1 +1,0 @@
-<kirby-tags-input :value="'design:photography:architecture'" :separator="':'" />
