--- src/Ui/Forms/Inputs/TagsInput/TagsInput.hbs
@@ -1,1 +1,0 @@
-<kirby-tags-input :value="['design', 'photography', 'architecture']" />
