--- src/Ui/Forms/Inputs/TagsInput/TagsInput.vue
@@ -1,235 +1,0 @@
-<template>
-  <draggable v-model="tags" :options="{disabled: !sortable}" class="kirby-tags-input">
-    <kirby-tag v-for="tag in tags"
-      :ref="tag"
-      :key="tag"
-      @blur.native="select(null)"
-      @focus.native="select(tag)"
-      @keydown.native.left="navigate('prev')"
-      @keydown.native.right="navigate('next')"
-      @dblclick.native="edit(tag)"
-      @remove="remove(tag)"
-      :removable="true">
-        {{ tag }}
-    </kirby-tag>
-    <span slot="footer" class="kirby-tags-input-element">
-      <kirby-autocomplete v-if="autocomplete"
-        ref="input"
-        :id="id"
-        :map="autocomplete.map"
-        :url="autocomplete.url"
-        :ignore="this.tags"
-        @keydown.native.left="leaveInput"
-        @keydown.native.delete="leaveInput"
-        @enter="function(value) { add(value) }"
-        @select="function(item) { add(item.value) }">
-      </kirby-autocomplete>
-      <input v-else :id="id" ref="input"
-        @keydown.enter="add($event.target.value)"
-        @keydown.left="leaveInput"
-        @keydown.delete="leaveInput">
-    </span>
-  </draggable>
-</template>
-
-<script>
-
-import draggable from 'vuedraggable'
-
-export default {
-  components: {
-    draggable
-  },
-  props: {
-    id: {},
-    value: {},
-    required: {},
-    separator: {
-      default: ','
-    },
-    autofocus: {},
-    autocomplete: {},
-    sortable: {
-      default: true
-    }
-  },
-  data: function() {
-
-    var tags = this.value || [];
-
-    if(typeof tags === 'string') {
-      var tags = tags.split(this.separator).map(tag => tag.trim());
-    }
-
-    return {
-      selected: null,
-      tags: tags
-    }
-
-  },
-  methods: {
-    focus () {
-      this.$refs.input.focus();
-    },
-    select (tag) {
-      this.selected = tag;
-    },
-    index (tag) {
-      return this.tags.indexOf(tag);
-    },
-    add (tag) {
-
-      var tag = tag.trim();
-
-      if(tag.length === 0) {
-        return;
-      }
-
-      if(this.index(tag) === -1) {
-        this.tags.push(tag);
-      }
-
-      if(this.autocomplete) {
-        this.$refs.input.close();
-        this.$refs.input.clear();
-      } else {
-        this.$refs.input.value = '';
-      }
-
-      this.$emit('input', this.tags);
-    },
-    edit (tag) {
-      this.$refs.input.value = tag;
-      this.$refs.input.select();
-      this.remove(tag);
-    },
-    remove (tag) {
-      var prev = this.get('prev');
-      var next = this.get('next');
-
-      this.tags.splice(this.index(tag), 1);
-
-      if(prev) {
-        prev.ref.focus();
-      } else if(next) {
-        this.$nextTick(() => {
-          var nextIndex  = this.tags.indexOf(next.tag);
-          var nextResult = this.get(nextIndex);
-          this.selected = nextResult.tag;
-          nextResult.ref.focus();
-        });
-      } else {
-        this.$refs.input.focus();
-      }
-
-      this.$emit('input', this.tags);
-    },
-    get (method) {
-
-      switch(method) {
-        case 'prev':
-        case 'next':
-          if(!this.selected) return;
-          var currIndex = this.index(this.selected);
-          var nextIndex = method === 'prev' ? (currIndex - 1) : (currIndex + 1);
-          break;
-        case 'first':
-          var nextIndex = 0;
-        case 'last':
-          var nextIndex = (this.tags.length - 1);
-          break;
-        default:
-          var nextIndex = method;
-      }
-
-      var nextTag = this.tags[nextIndex];
-      var nextRef = this.$refs[nextTag];
-
-      if(nextRef && nextRef[0]) {
-        return {
-          ref: nextRef[0],
-          tag: nextTag,
-          index: nextIndex
-        }
-      } else {
-        return false;
-      }
-
-    },
-    navigate (method) {
-
-      var result = this.get(method);
-
-      if(result) {
-        result.ref.focus();
-        this.selected = result.tag;
-      } else if(method === 'next') {
-        this.$refs.input.focus();
-        this.selected = null;
-      }
-
-    },
-    leaveInput (e) {
-
-      var position = e.target.selectionStart;
-
-      if(position === 0) {
-        this.navigate('last');
-        e.target.blur();
-      }
-
-    }
-  }
-}
-
-</script>
-
-<style lang="scss">
-
-  .kirby-tags-input {
-    display: flex;
-    flex-wrap: wrap;
-    align-items: baseline;
-
-    [dir="ltr"] & {
-      padding: 4px 0 0 4px;
-    }
-    [dir="rtl"] & {
-      padding: 4px 4px 0 0;
-    }
-  }
-  .kirby-tags-input .kirby-tag {
-    margin-bottom: 4px;
-
-    [dir="ltr"] & {
-      margin-right: 4px;
-    }
-    [dir="rtl"] & {
-      margin-left: 4px;
-    }
-  }
-  .kirby-tags-input .kirby-tag.sortable-ghost {
-    opacity: .2;
-  }
-  .kirby-tags-input-element input {
-    display: inline-block;
-    min-width: 4rem;
-    padding: .4rem;
-    flex-grow: 1;
-    font: inherit;
-    border: 0;
-    outline: 0;
-    line-height: 1;
-    outline: 0;
-    margin-bottom: 4px;
-    border-radius: 3px;
-
-    [dir="ltr"] & {
-      margin-right: 4px;
-    }
-    [dir="rtl"] & {
-      margin-left: 4px;
-    }
-  }
-
-</style>
