+++ src/Ui/Forms/Inputs/TextInput/TextInput.hbs
@@ -1,5 +1,4 @@
 <kirby-text-input
   name="text"
-  placeholder="Write something here…"
-  :required="true">
+  placeholder="Write something here…">
 </kirby-text-input>
